+++ resources/views/partials/bootstrap-5-alert-message.blade.php
@@ -1,8 +1,7 @@
-
 <div class="alert alert-{!! ($type === 'error' ? 'danger' : $type) !!}" role="alert">
-    @if (!empty($title))
-        <strong>{{ $title }}</strong><br>
+    @if (! empty($title))
+        <h4 class="alert-heading">{!! $title !!}</h4>
     @endif
 
-    {{ $message }}
+    {!! $message !!}
 </div>