+++ resources/views/partials/pnotify-5-message.blade.php
@@ -1,0 +1,16 @@
+@php
+    $PNotifyParams = [
+        'type' => $type === 'warning' ? 'notice' : $type,
+        'text' => $message,
+    ];
+
+    if ($title) {
+        $PNotifyParams['title'] = $title;
+    }
+
+    $PNotifyParams = json_encode($PNotifyParams);
+@endphp
+
+<script>
+    PNotify.alert({!! $PNotifyParams !!});
+</script>